+++ CHANGES.rst
@@ -45,6 +45,8 @@ Improvements
 - Include non-speaker authors in the timetable export API (:issue:`5412`, :pr:`5738`)
 - Add setting to force track selection when accepting abstracts (:pr:`5771`)
 - Log detailed changes when editing contributions (:pr:`5777`)
+- Allow managers to ignore required field restrictions in registration forms
+  (:issue:`5644`, :pr:`5682`, thanks :user:`kewisch`)
 
 Bugfixes
 ^^^^^^^^